% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{as.data.frame.QuantileForecastMatrix}
\alias{as.data.frame.QuantileForecastMatrix}
\title{Translate from a QuantileForecastMatrix representation of forecasts to a
tidy data frame}
\usage{
\method{as.data.frame}{QuantileForecastMatrix}(qfm)
}
\arguments{
\item{qfm}{wide matrix representation of forecasts}
}
\value{
tidy data frame of forecasts
}
\description{
Translate from a QuantileForecastMatrix representation of forecasts to a
tidy data frame
}
