% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline_forecasts}
\alias{baseline_forecasts}
\title{Forecasts from a simple baseline model}
\format{
A tibble with a row per week end date
}
\usage{
baseline_forecasts
}
\description{
Forecasts from a simple baseline model
}
\keyword{datasets}
