% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{build_covid_ensemble_from_zoltar}
\alias{build_covid_ensemble_from_zoltar}
\title{Download covid forecasts from zoltar and fit one ensemble}
\usage{
build_covid_ensemble_from_zoltar(
  candidate_model_abbreviations_to_include,
  targets,
  forecast_week_end_date,
  timezero_window_size = 1,
  window_size,
  ensemble_method,
  project_url = "https://www.zoltardata.com/api/project/44/",
  required_quantiles,
  return_eligibility = TRUE
)
}
\arguments{
\item{candidate_model_abbreviations_to_include}{List of model abbreviations
for models that may be included in ensemble forecast}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death')}

\item{forecast_week_end_date}{date of a s}

\item{timezero_window_size}{The number of days back to go.  A window size of
0 will retrieve only forecasts submitted on the \code{last_timezero} date.}

\item{window_size}{size of window}

\item{ensemble_method}{name of method to use: 'ew', 'convex',
'unconstrained', or 'rescaled_convex'}

\item{project_url}{zoltar project url}

\item{required_quantiles}{numeric vector of quantiles component models are
required to have submitted}

\item{return_eligibility}{if TRUE, return model eligibility}
}
\value{
data frame with ensemble forecasts by location
}
\description{
Download covid forecasts from zoltar and fit one ensemble
}
