% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eligibility.R
\name{calc_baseline_check}
\alias{calc_baseline_check}
\title{Compute check of whether predictive performance is comparable to a baseline
model}
\usage{
calc_baseline_check(
  qfm,
  observed_by_location_target_end_date,
  baseline_tol = 1
)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{observed_by_location_target_end_date}{data frame of observed values}

\item{baseline_tol}{numeric; for baseline check, model's mean WIS for
forecasts within the window size must be at most baseline_tol times the
mean wis for the baseline model on the corresponding forecasts}
}
\value{
data frame with a row for each combination of
location and model, and a character column called
'baseline_eligibility' with entry 'windowed wis greater than baseline'
if the mean wis for forecasts made by this model in the past window weeks
is greater than the mean wis for the corresponding forecasts from a
baseline model
}
\description{
Compute check of whether predictive performance is comparable to a baseline
model
}
