% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eligibility.R
\name{calc_q10_check}
\alias{calc_q10_check}
\title{Compute forecast missingness for each combination of unit and model}
\usage{
calc_q10_check(qfm, observed, model_id_name)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{observed}{data frame of observed values}

\item{model_id_name}{character name of column in forecast matrix col_index
specifying column}
}
\value{
data frame with a row for each combination of
unit, forecast week end date, and model and a logical column called
'any_missing' with entry TRUE if any forecasts were missing across all
quantiles and targets
}
\description{
Compute forecast missingness for each combination of unit and model
}
