% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_calculations.R
\name{calc_qra}
\alias{calc_qra}
\title{Calculate quantile regression average of forecast quantiles from component
models.}
\usage{
calc_qra(forecast_df, weight_df, model_col, id_cols, quantile_cols)
}
\arguments{
\item{forecast_df}{data frame of forecasts from component models. Each row
corresponds to a single forecast.  At minimum, must contain:
\itemize{
\item a column with a model identifier
\item a separate column for each quantile to be aggregated
Optionally, also contains columns uniquely identifying the forecast such as
a location, time, and horizon.
}}

\item{weights_df}{data frame of weights assigned to each model, including an
intercept term}
}
\description{
Calculate quantile regression average of forecast quantiles from component
models.
}
