% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qra.R
\name{calc_qra_forecast}
\alias{calc_qra_forecast}
\title{An internal low level function to calculate quantile regression averaged
forecasts from suitably defined matrices of forecasts and weights.}
\usage{
calc_qra_forecast(qfm, intercepts, weights)
}
\description{
An internal low level function to calculate quantile regression averaged
forecasts from suitably defined matrices of forecasts and weights.
}
