% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_calculations.R
\name{calc_qra_from_forecast_matrix}
\alias{calc_qra_from_forecast_matrix}
\title{An internal low level function to calculate quantile regression averaged
forecasts from suitably defined matrices of forecasts and weights.}
\usage{
calc_qra_from_forecast_matrix(forecast_matrix, intercepts, weights)
}
\description{
An internal low level function to calculate quantile regression averaged
forecasts from suitably defined matrices of forecasts and weights.
}
