% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_management.R
\name{calc_relative_horizon}
\alias{calc_relative_horizon}
\title{Calculate the effective horizon of a forecast relative to the
\code{forecast_week_end_date}}
\usage{
calc_relative_horizon(forecast_week_end_date, target_end_date, target)
}
\arguments{
\item{forecast_week_end_date}{date or vector of dates of the
same length as target_end_date, relative to which horizons should be
calculated}

\item{target_end_date}{vector of dates defining dates targeted by forecasts,
of same length as target}

\item{target}{vector of strings specifying targets, e.g.
"1 wk ahead inc case", of same length as target_end_date}
}
\value{
vector of horizons of forecasts relative to the
forecast_week_end_date, in units appropriate to the target scale (wk or day)
}
\description{
Calculate the effective horizon of a forecast relative to the
\code{forecast_week_end_date}
}
