% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_management.R
\name{calc_relative_target}
\alias{calc_relative_target}
\title{Calculate the effective target relative to the forecast_week_end_date}
\usage{
calc_relative_target(forecast_week_end_date, target_end_date, target)
}
\arguments{
\item{forecast_week_end_date}{date or vector of dates of the
same length as target_end_date, relative to which horizons should be
calculated}

\item{target_end_date}{vector of dates defining dates targeted by forecasts}

\item{target}{vector of strings specifying targets, e.g.
"1 wk ahead inc case"}
}
\value{
vector of strings defining the target relative to the
forecast_week_end_date
}
\description{
Calculate the effective target relative to the forecast_week_end_date
}
