% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective_utils.R
\name{calc_retrospective_ensemble_scores}
\alias{calc_retrospective_ensemble_scores}
\title{Calculate scores of retrospective ensemble forecasts}
\usage{
calc_retrospective_ensemble_scores(
  submissions_root,
  forecast_dates,
  all_locations,
  spatial_scales = c("national", "state", "state_national", "county"),
  response_vars = NULL,
  truth_as_of = NULL
)
}
\arguments{
\item{truth_as_of}{Date or string in format "yyyy-mm-dd" specifying
as_of date to use for truth}

\item{forecast_files}{character vector of paths to retrospective ensemble
forecast files}
}
\value{
data frame of scores
}
\description{
Calculate scores of retrospective ensemble forecasts
}
