% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eligibility.R
\name{calc_sd_check}
\alias{calc_sd_check}
\title{Compute check of whether mean of forecast medians in \code{lookahead_window}
is more than \code{num_sd} sample standard deviations below the sample mean of observations
in \code{lookback_window}. If \code{sd_check_table_path} is supplied, a table with the
excluded location-model pairs is saved there, and if \code{sd_check_plot_path} is supplied,
plots illustration the exclusion criteria are saved there.}
\usage{
calc_sd_check(
  qfm,
  observed_by_location_target_end_date,
  sd_check_table_path = NULL,
  sd_check_plot_path = NULL,
  lookback_inside = 0,
  lookback_outside = 13,
  lookahead_inside = 1,
  lookahead_outside = 7,
  num_sd = 4,
  show_stats = FALSE
)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{observed_by_location_target_end_date}{data frame of observed values}

\item{sd_check_table_path}{}

\item{sd_check_plot_path}{}

\item{lookback_inside}{number of days to be excluded going back from time zero from mean
calculation.  Intended to be used to adjust for unreliable recent reporting.}

\item{lookback_outside}{number of days beyond day zero to look back.  Should be
1 less than desired window size if day zero is include in window.}

\item{lookahead_inside}{}

\item{lookahead_outside}{}

\item{num_sd}{number of standard deviations to use for cutoff}
}
\value{
data frame with a row for each combination of
location and model and a logical column called
'eligibility' with entry "more than four SDs below 2-week mean of
hospitalization point forecasts" if check fails
}
\description{
Compute check of whether mean of forecast medians in \code{lookahead_window}
is more than \code{num_sd} sample standard deviations below the sample mean of observations
in \code{lookback_window}. If \code{sd_check_table_path} is supplied, a table with the
excluded location-model pairs is saved there, and if \code{sd_check_plot_path} is supplied,
plots illustration the exclusion criteria are saved there.
}
