% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_management.R
\name{calc_target_string_end_date}
\alias{calc_target_string_end_date}
\title{Calculate the target end date by parsing a target specification string and
adding appropriate integer to \code{start_date}.}
\usage{
calc_target_string_end_date(start_date, target)
}
\arguments{
\item{start_date}{a Date object}

\item{target}{string specification of forecast target}
}
\value{
target end date
}
\description{
Calculate the target end date by parsing a target specification string and
adding appropriate integer to \code{start_date}.
}
