% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidEnsembles-package.R
\docType{package}
\name{covidEnsembles}
\alias{covidEnsembles}
\alias{covidEnsembles-package}
\title{covidEnsembles - Utility functions for CDC forecasts}
\description{
Building QRA ensembles for covid
}
\author{
Evan L. Ray \href{mailto:eray@mtholyoke.edu}{eray@mtholyoke.edu}
}
