% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{do_zoltar_query}
\alias{do_zoltar_query}
\title{Utility function to construct and execute a Zoltar query for all quantile
forecasts submitted to a project within a specified number of days of a
given date.}
\usage{
do_zoltar_query(
  model_abbrs,
  last_timezero,
  timezero_window_size,
  targets,
  zoltar_connection,
  project_url,
  verbose
)
}
\arguments{
\item{model_abbrs}{Character vector of model abbreviations}

\item{last_timezero}{The last timezero date of forecasts to retrieve}

\item{timezero_window_size}{The number of days back to go.  A window size of
0 will retrieve only forecasts submitted on the \code{last_timezero} date.}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death')}

\item{zoltar_connection}{zoltar connection}

\item{project_url}{zoltar project url}

\item{verbose}{if TRUE, print messages on job status poll}
}
\value{
data frame of forecasts
}
\description{
Utility function to construct and execute a Zoltar query for all quantile
forecasts submitted to a project within a specified number of days of a
given date.
}
