% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eligibility.R
\name{drop_ineligible_forecasts}
\alias{drop_ineligible_forecasts}
\title{drop ineligible forecasts}
\usage{
drop_ineligible_forecasts(forecasts, eligibility)
}
\arguments{
\item{forecasts}{data frame of forecasts}

\item{model_eligibility}{data frame with model eligibility results as
returned by calc_model_eligibility_for_ensemble}
}
\value{
data frame of forecasts that are eligible
}
\description{
drop ineligible forecasts
}
