% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{estimate_qra_ew}
\alias{estimate_qra_ew}
\title{Estimate Quantile Regression Averaging model with equal weights}
\usage{
estimate_qra_ew(qfm_train, ...)
}
\arguments{
\item{qfm_train}{}
}
\value{
object of class qra_fit
}
\description{
Estimate Quantile Regression Averaging model with equal weights
}
