% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{estimate_qra_quantgen}
\alias{estimate_qra_quantgen}
\title{Estimate qra model using quantgen package as backend}
\usage{
estimate_qra_quantgen(
  qfm_train,
  y_train,
  qfm_test,
  intercept,
  constraint,
  quantile_groups,
  noncross = "constrain"
)
}
\arguments{
\item{qfm_train}{QuantileForecastMatrix with training set predictions}

\item{y_train}{numeric vector of responses for training set}

\item{qfm_test}{QuantileForecastMatrix with test set predictions}

\item{intercept}{logical specifying whether an intercept is included}

\item{constraint}{character specifying constraints on parameters; 'convex',
'positive' or 'unconstrained'}

\item{quantile_groups}{Vector of group labels for quantiles, having the same
length as the number of quantiles.  Common labels indicate that the ensemble
weights for the corresponding quantile levels should be tied together.
Default is rep(1,length(quantiles)), which means that a common set of
ensemble weights should be used across all levels.  This is the argument
\code{tau_groups} for \code{quantgen::quantile_ensemble}}

\item{noncross}{string specifying approach to handling quantile noncrossing:
one of "constrain" or "sort". "constrain" means estimation is done subject
to constraints ruling out quantile crossing.  "sort" means no such
constraints are imposed during estimation, but the resulting forecasts are
sorted.}
}
\value{
object of class qra_fit
}
\description{
Estimate qra model using quantgen package as backend
}
