% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_zoltar_forecasts.R
\name{get_all_project_forecasts}
\alias{get_all_project_forecasts}
\title{Download all forecasts for a project and assemble into a data frame}
\usage{
get_all_project_forecasts(zoltar_connection, project_url, ...)
}
\arguments{
\item{zoltar_connection}{A 'ZoltarConnection' object as returned by new_connection}

\item{project_url}{URL of a project in zoltar_connection's projects}

\item{...}{conditions for filtering the results; will be passed to dplyr::filter}
}
\value{
data frame with forecasts meeting the specified filter conditions
}
\description{
Download all forecasts for a project and assemble into a data frame
}
