% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{get_all_wis_components}
\alias{get_all_wis_components}
\title{compute wis and contributions to wis from each interval level}
\usage{
get_all_wis_components(qfm, observed_by_location_target_end_date)
}
\arguments{
\item{qfm}{QuantileForecastMatrix with predictions to score}

\item{observed_by_location_target_end_date}{data frame of observed values}
}
\value{
data frame with column per component of wis
}
\description{
compute wis and contributions to wis from each interval level
}
