% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{get_by_location_group_ensemble_fits_and_predictions}
\alias{get_by_location_group_ensemble_fits_and_predictions}
\title{Calculate ensemble fits separately by location group}
\usage{
get_by_location_group_ensemble_fits_and_predictions(
  forecasts,
  observed_by_location_target_end_date,
  forecast_week_end_date,
  window_size,
  ensemble_method = c("ew", "convex", "unconstrained"),
  return_all = FALSE,
  return_eligibility = TRUE
)
}
\arguments{
\item{forecasts}{data frame with columns 'model', 'location',
'forecast_week_end_date', 'target', 'quantile', and 'value'}

\item{observed_by_location_target_end_date}{data frame with columns
'location', 'base_target', 'target_end_date', and 'observed'}

\item{forecast_week_end_date}{Date object: date of the saturday for the end
of the forecast week; week-ahead targets are with respect to this date}

\item{window_size}{size of window}

\item{ensemble_method}{name of method to use: 'ew', 'convex', or
'unconstrained'}

\item{return_all}{if TRUE, return all quantities; if FALSE, return only some
useful summaries}

\item{return_eligibility}{if TRUE, return model eligibility}
}
\value{
tibble or data frame with ensemble fits and results
}
\description{
Calculate ensemble fits separately by location group
}
