% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{get_candidate_models}
\alias{get_candidate_models}
\title{Get a list of candidate models with specified model designations}
\usage{
get_candidate_models(
  submissions_root,
  include_designations = c("primary", "secondary", "proposed"),
  include_COVIDhub_ensemble = FALSE,
  include_COVIDhub_baseline = TRUE
)
}
\arguments{
\item{submissions_root}{path to the data-processed folder of the
covid19-forecast-hub repository}

\item{include_designations}{character vector of model designations to
include: "primary", "secondary", "proposed", and/or "other"}

\item{include_COVIDhub_ensemble}{logical: if TRUE, whether or not the
COVIDhub-ensemble model is included depends on whether it falls within one
of the specified \code{include_designations}; if FALSE, it will never be
included}

\item{include_COVIDhub_baseline}{logical: if TRUE, whether or not the
COVIDhub-baseline model is included depends on whether it falls within one
of the specified \code{include_designations}; if FALSE, it will never be
included}
}
\value{
character vector of model abbreviations
}
\description{
Get a list of candidate models with specified model designations
}
