% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth.R
\name{get_observed_by_location_target_end_date}
\alias{get_observed_by_location_target_end_date}
\title{get observed cases and/or deaths}
\usage{
get_observed_by_location_target_end_date(
  issue_date,
  targets,
  spatial_resolution
)
}
\arguments{
\item{issue_date}{character issue date (i.e. report date) to use for
constructing truths in format 'yyyy-mm-dd'}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death')}

\item{spatial_resolution}{character vector specifying spatial unit types to
include: 'county', 'state' and/or 'national'}
}
\value{
data frame with columns location, base_target, target_end_date, and
observed
}
\description{
get observed cases and/or deaths
}
