% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_zoltar_forecasts.R
\name{get_one_project_forecast}
\alias{get_one_project_forecast}
\title{Download a single forecast for a project and assemble into a data frame}
\usage{
get_one_project_forecast(zoltar_connection, forecast_url, ...)
}
\arguments{
\item{zoltar_connection}{A 'ZoltarConnection' object as returned by new_connection}

\item{...}{conditions for filtering the results; will be passed to dplyr::filter}

\item{project_url}{URL of a project in zoltar_connection's projects}
}
\value{
data frame with forecasts meeting the specified filter conditions
}
\description{
Download a single forecast for a project and assemble into a data frame
}
