% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{get_rescaled_ensemble_fits_and_predictions}
\alias{get_rescaled_ensemble_fits_and_predictions}
\title{Calculate rescaled convex ensemble fit, renormalizing in case of missing models}
\usage{
get_rescaled_ensemble_fits_and_predictions(
  forecasts,
  observed_by_location_target_end_date,
  forecast_week_end_date,
  window_size,
  do_q10_check,
  do_nondecreasing_quantile_check,
  do_baseline_check = do_baseline_check,
  do_sd_check,
  sd_check_table_path = NULL,
  sd_check_plot_path = NULL,
  baseline_tol = 1.2,
  manual_eligibility_adjust,
  return_all = FALSE,
  return_eligibility = TRUE
)
}
\arguments{
\item{forecasts}{data frame with columns 'model', 'location',
'forecast_week_end_date', 'target', 'quantile', and 'value'}

\item{observed_by_location_target_end_date}{data frame with columns
'location', 'base_target', 'target_end_date', and 'observed'}

\item{forecast_week_end_date}{Date object: date of the saturday for the end
of the forecast week; week-ahead targets are with respect to this date}

\item{window_size}{size of window}
}
\value{
tibble or data frame with ensemble fits and results
}
\description{
Calculate rescaled convex ensemble fit, renormalizing in case of missing models
}
