% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{get_rescaled_ensemble_fits_and_predictions_one_week_and_window_size}
\alias{get_rescaled_ensemble_fits_and_predictions_one_week_and_window_size}
\title{Calculate ensemble fits separately by location group}
\usage{
get_rescaled_ensemble_fits_and_predictions_one_week_and_window_size(
  forecasts,
  observed_by_location_target_end_date,
  forecast_week_end_date,
  window_size
)
}
\arguments{
\item{forecasts}{data frame with columns 'model', 'location',
'forecast_week_end_date', 'target', 'quantile', and 'value'}

\item{observed_by_location_target_end_date}{data frame with columns
'location', 'base_target', 'target_end_date', and 'observed'}

\item{forecast_week_end_date}{Date object: date of the saturday for the end
of the forecast week; week-ahead targets are with respect to this date}

\item{window_size}{size of window}
}
\value{
tibble or data frame with ensemble fits and results
}
\description{
Calculate ensemble fits separately by location group
}
