% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{grid_search_rel_wis_weighted_median}
\alias{grid_search_rel_wis_weighted_median}
\title{Conduct a grid search by evaluating the objective function at all parameter
values in the supplied parameter grid and selecting the parameter value
with lowest function value.  Works only with one-dimensional parameters.}
\usage{
grid_search_rel_wis_weighted_median(
  par_grid,
  fn,
  model_constructor,
  qfm_train,
  y_train,
  quantile_groups
)
}
\arguments{
\item{par_grid}{numeric vector of parameter values to evaluate}

\item{fn}{objective function to minimize}

\item{model_constructor}{function accepts a real-valued parameter and
returns a model of class qra_fit}

\item{qfm_train}{QuantileForecastMatrix with training set predictions from
component models}

\item{y_train}{numeric vector of responses for training set}

\item{quantile_groups}{Vector of group labels for quantiles, having the same
length as the number of quantiles.  Common labels indicate that the ensemble
weights for the corresponding quantile levels should be tied together.
Default is rep(1,length(quantiles)), which means that a common set of
ensemble weights should be used across all levels.  This is the argument
\code{tau_groups} for \code{quantgen::quantile_ensemble}, and may only be supplied if
\verb{backend = 'quantgen}}
}
\value{
single parameter value with lowest function value
}
\description{
Conduct a grid search by evaluating the objective function at all parameter
values in the supplied parameter grid and selecting the parameter value
with lowest function value.  Works only with one-dimensional parameters.
}
