% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{grid_search_rel_wis_weights}
\alias{grid_search_rel_wis_weights}
\title{Conduct a grid search by evaluating the objective function at all parameter
values in the supplied parameter grid and selecting the parameter value
with lowest function value.  Works only with one-dimensional parameters.}
\usage{
grid_search_rel_wis_weights(
  par_grid,
  fn,
  model_constructor,
  qfm_train,
  y_train,
  max_weight
)
}
\arguments{
\item{par_grid}{numeric vector of parameter values to evaluate}

\item{fn}{objective function to minimize}

\item{model_constructor}{function accepts a real-valued parameter and
returns a model of class qra_fit}

\item{qfm_train}{QuantileForecastMatrix with training set predictions from
component models}

\item{y_train}{numeric vector of responses for training set}

\item{max_weight}{numeric value for maximum weight}
}
\value{
single parameter value with lowest function value
}
\description{
Conduct a grid search by evaluating the objective function at all parameter
values in the supplied parameter grid and selecting the parameter value
with lowest function value.  Works only with one-dimensional parameters.
}
