% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_truth.R
\name{historical_truths}
\alias{historical_truths}
\title{Construct table of truths for weekly incident and cumulative deaths at
horizons one though 6}
\usage{
historical_truths(
  issue_date = NULL,
  spatial_resolution = "state",
  temporal_resolution = "weekly"
)
}
\arguments{
\item{issue_date}{issue date (i.e. report date) to use for constructing
truths}

\item{spatial_resolution}{character vector specifying spatial unit types to
include: 'state' and/or 'national'}

\item{temporal_resolution}{character vector specifying temporal resolution
to include: currently only 'weekly' is supported}
}
\value{
data frame with
}
\description{
Construct table of truths for weekly incident and cumulative deaths at
horizons one though 6
}
