% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{impute_missing_per_quantile}
\alias{impute_missing_per_quantile}
\title{Impute missing values for each quantile level in a quantile forecast matrix
It is assumed that in each row, all quantiles for a given model are either
missing or available.}
\usage{
impute_missing_per_quantile(qfm, impute_method = "mean")
}
\arguments{
\item{qfm}{a QuantileForecastMatrix}

\item{impute_method}{character string specifying method for imputing;
currently only 'mean' is supported}
}
\value{
list of two items:
\enumerate{
\item 'qfm_imputed' the input QuantileForecastMatrix object with missing values
imputed
\item 'weight_transfer' a square matrix of dimension equal to the number of
unique models in qfm.  Entry \link{i, j} is the proportion of imputed observations
for model j that are attributable to model i.
}
}
\description{
Impute missing values for each quantile level in a quantile forecast matrix
It is assumed that in each row, all quantiles for a given model are either
missing or available.
}
