% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{init_par_constructor_rel_wis_weighted_mean}
\alias{init_par_constructor_rel_wis_weighted_mean}
\title{Initial parameter constructor for rel_wis_weighted_mean approach}
\usage{
init_par_constructor_rel_wis_weighted_mean(qfm_train, ...)
}
\arguments{
\item{qfm_train}{QuantileForecastMatrix with training set predictions from
component models}

\item{...}{mop up other arguments that are ignored}
}
\value{
vector of real-valued initial values for parameters
}
\description{
Initial parameter constructor for rel_wis_weighted_mean approach
}
