% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{init_par_constructor_unconstrained_per_model}
\alias{init_par_constructor_unconstrained_per_model}
\title{Initial parameter constructor for unconstrained_per_model approach}
\usage{
init_par_constructor_unconstrained_per_model(qfm_train, ...)
}
\arguments{
\item{qfm_train}{QuantileForecastMatrix with training set predictions from
component models}

\item{...}{mop up other arguments that are ignored}
}
\value{
vector of real-valued initial values for parameters
}
\description{
Initial parameter constructor for unconstrained_per_model approach
}
