% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{is.median_qra_fit}
\alias{is.median_qra_fit}
\title{Check if object is of class median_qra_fit}
\usage{
is.median_qra_fit(object)
}
\arguments{
\item{object}{an object that may be a median_qra_fit object}
}
\value{
boolean; whether object is inherits median_qra_fit class
}
\description{
Check if object is of class median_qra_fit
}
