% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{is.qenspy_qra_fit}
\alias{is.qenspy_qra_fit}
\title{Check if object is of class qenspy_qra_fit}
\usage{
is.qenspy_qra_fit(object)
}
\arguments{
\item{object}{an object that may be a qenspy_qra_fit object}
}
\value{
boolean; whether object is inherits qenspy_qra_fit class
}
\description{
Check if object is of class qenspy_qra_fit
}
