% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{is.QuantileForecastMatrix}
\alias{is.QuantileForecastMatrix}
\title{Check if object is of class QuantileForecastMatrix}
\usage{
is.QuantileForecastMatrix(qfm)
}
\arguments{
\item{qfm}{an object that may be a QuantileForecastMatrix}
}
\value{
boolean; whether object is inherits QuantileForecastMatrix class
}
\description{
Check if object is of class QuantileForecastMatrix
}
