% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{is.weighted_median_qra_fit}
\alias{is.weighted_median_qra_fit}
\title{Check if object is of class weighted_median_qra_fit}
\usage{
is.weighted_median_qra_fit(object)
}
\arguments{
\item{object}{an object that may be a weighted_median_qra_fit object}
}
\value{
boolean; whether object is inherits weighted_median_qra_fit class
}
\description{
Check if object is of class weighted_median_qra_fit
}
