% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jhu_data}
\alias{jhu_data}
\title{JHU data for the US and its states and territories}
\format{
A tibble with a row per issue date
}
\usage{
jhu_data
}
\description{
JHU data for the US and its states and territories
}
\keyword{datasets}
