% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{load_covid_forecasts}
\alias{load_covid_forecasts}
\title{load covid forecasts from local file}
\usage{
load_covid_forecasts(
  model_abbrs,
  last_timezero,
  timezero_window_size,
  targets,
  submissions_root
)
}
\arguments{
\item{model_abbrs}{Character vector of model abbreviations}

\item{last_timezero}{The last timezero date of forecasts to retrieve}

\item{timezero_window_size}{The number of days back to go.  A window size of
0 will retrieve only forecasts submitted on the \code{last_timezero} date.}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death')}

\item{submissions_root}{path to the data-processed folder of the
covid19-forecast-hub repository}
}
\value{
data frame with columns model, timezero, location, target, quantile,
value
}
\description{
load covid forecasts from local file
}
