% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{load_covid_forecasts_relative_horizon}
\alias{load_covid_forecasts_relative_horizon}
\title{Load multiple models' last forecasts from multiple weeks and
adjust the targets for daily forecasts to be relative to the
Monday on or after submission.}
\usage{
load_covid_forecasts_relative_horizon(
  hub = "US",
  source = "local_hub_repo",
  hub_repo_path,
  data_processed_subpath = "data-processed/",
  monday_dates,
  as_of,
  model_abbrs,
  timezero_window_size,
  locations,
  targets,
  max_horizon,
  required_quantiles
)
}
\arguments{
\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are "US" and "ECDC"}

\item{source}{string specifying where forecasts will be loaded from: either
"local_hub_repo" or "zoltar"}

\item{hub_repo_path}{path to local clone of the reichlab/covid19-forecast-hub
repository to be used when \code{source} is \code{local_hub_repo}}

\item{monday_dates}{Date vector of Mondays that are submission deadlines}

\item{as_of}{"as_of" date passed to zoltar
Ignored if \code{source} is \code{"local_hub_repo"}.}

\item{model_abbrs}{Character vector of model abbreviations}

\item{timezero_window_size}{The number of days back to go.  A window size of
0 will retrieve only forecasts submitted on the \code{last_timezero} date.}

\item{locations}{character vector of locations; FIPS codes}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death')}

\item{max_horizon}{maximum horizon of forecasts to return, where horizon is relative to
the reference dates for the forecasts (Saturdays for forecasts of weekly targets or Mondays for
daily targets)}

\item{required_quantiles}{numeric vector of quantiles component models are
required to have submitted}
}
\value{
data frame with ensemble forecasts by location
}
\description{
Load multiple models' last forecasts from multiple weeks and
adjust the targets for daily forecasts to be relative to the
Monday on or after submission.
}
