% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_loading_code.R
\name{load_covid_forecasts_relative_horizon_old}
\alias{load_covid_forecasts_relative_horizon_old}
\title{Load forecasts from multiple models over multiple weeks, keep only the last
forecast from each model each week, and adjust the targets for daily
forecasts to be relative to the Monday on or after submission.}
\usage{
load_covid_forecasts_relative_horizon_old(
  monday_dates,
  model_abbrs,
  timezero_window_size,
  locations,
  targets,
  horizon,
  required_quantiles,
  submissions_root,
  include_null_point_forecasts = FALSE,
  keep_last = TRUE
)
}
\arguments{
\item{monday_dates}{Date vector of Mondays that are submission deadlines}

\item{model_abbrs}{Character vector of model abbreviations}

\item{timezero_window_size}{The number of days back to go.  A window size of
0 will retrieve only forecasts submitted on the \code{last_timezero} date.}

\item{locations}{character vector of locations; FIPS codes}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death')}

\item{horizon}{maximum horizon relative to the monday dates that should be
retained}

\item{required_quantiles}{numeric vector of quantiles component models are
required to have submitted}

\item{submissions_root}{path to the data-processed folder of the
covid19-forecast-hub repository}

\item{include_null_point_forecasts}{boolean; if TRUE, quantile forecasts are
augmented with NA point forecasts from all models that submitted point
forecasts}

\item{keep_last}{logical; if TRUE, keep only the last identified file
within each week. if FALSE, keep all files.}
}
\value{
data frame with columns model, timezero, location, target, quantile,
value
}
\description{
Load forecasts from multiple models over multiple weeks, keep only the last
forecast from each model each week, and adjust the targets for daily
forecasts to be relative to the Monday on or after submission.
}
