% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{model_constructor_rel_wis_weighted_mean}
\alias{model_constructor_rel_wis_weighted_mean}
\title{Model constructor for rel_wis_weighted_mean approach}
\usage{
model_constructor_rel_wis_weighted_mean(
  par,
  qfm_train,
  y_train,
  rel_wis = calc_relative_wis(y_train, qfm_train)
)
}
\arguments{
\item{par}{vector of real numbers}

\item{qfm_train}{object of class QuantileForecastMatrix}
}
\value{
object of class qra_fit
}
\description{
Model constructor for rel_wis_weighted_mean approach
}
