% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{model_constructor_rescaled_convex_per_model}
\alias{model_constructor_rescaled_convex_per_model}
\title{Model constructor for rescaled_convex_per_model approach}
\usage{
model_constructor_rescaled_convex_per_model(par, qfm_train)
}
\arguments{
\item{par}{vector of real numbers}

\item{qfm_train}{object of class QuantileForecastMatrix}
}
\value{
object of class rescaled_qra_fit
}
\description{
Model constructor for rescaled_convex_per_model approach
}
