% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{new_median_qra_fit}
\alias{new_median_qra_fit}
\title{Create a median_qra_fit object}
\usage{
new_median_qra_fit(...)
}
\arguments{
\item{...}{any arguments are ignored}
}
\value{
median_qra_fit object
}
\description{
Create a median_qra_fit object
}
