% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{new_qenspy_qra_fit}
\alias{new_qenspy_qra_fit}
\title{Create a qenspy_qra_fit object}
\usage{
new_qenspy_qra_fit(
  param_vec,
  M,
  model_names = rep("", M),
  quantile_levels,
  quantile_groups,
  combine_method,
  loss_trace = NULL
)
}
\arguments{
\item{param_vec}{numeric vector of "raw" parameter estimates}

\item{M}{integer number of models}

\item{model_names}{name for each model}

\item{quantile_levels}{numeric vector of quantile levels}

\item{quantile_groups}{number of quantile groups}

\item{combine_method}{string specifying combination method;
either "convex_mean" or "convex_median"}

\item{loss_trace}{numeric vector of loss trace per iteration of estimation}
}
\value{
qenspy_qra_fit object
}
\description{
Create a qenspy_qra_fit object
}
