% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{new_qra_fit}
\alias{new_qra_fit}
\title{Create a qra_fit object}
\usage{
new_qra_fit(parameters, convex)
}
\arguments{
\item{parameters}{named list with two components:
\itemize{
\item 'intercept': data frame optionally with a column for quantile values,
and corresponding intercepts
\item 'coefficients': data frame of models, optionally quantile values,
and corresponding model coefficients
}}

\item{convex}{logical; if TRUE, predictions are done by passing weights
through a softmax transformation; otherwise, no constraint is imposed.}
}
\value{
qra_fit object
}
\description{
Create a qra_fit object
}
