% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{new_QuantileForecastMatrix}
\alias{new_QuantileForecastMatrix}
\title{Create a QuantileForecastMatrix object}
\usage{
new_QuantileForecastMatrix(
  qfm,
  row_index,
  col_index,
  quantile_name_col,
  quantile_value_col
)
}
\arguments{
\item{qfm}{a matrix of forecasts.  Distinct combinations of unit,}

\item{row_index}{data frame giving unique combination of identifiers such as
location, forecast date, and target for each row of \code{qfm}}

\item{col_index}{data frame giving the combination of model and quantile
corresponding to each column of \code{qfm}}

\item{quantile_name_col}{name of column for quantile probability in long
format data frame representation of forecasts}

\item{quantile_value_col}{name of column for quantile value in long format
data frame representation of forecasts}
}
\value{
QuantileForecastMatrix object
}
\description{
Create a QuantileForecastMatrix object
}
