% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{new_weighted_median_qra_fit}
\alias{new_weighted_median_qra_fit}
\title{Create a weighted_median_qra_fit object}
\usage{
new_weighted_median_qra_fit(parameters)
}
\arguments{
\item{parameters}{named list with two components:
\itemize{
\item 'intercept': data frame optionally with a column for quantile values,
and corresponding intercepts
\item 'coefficients': data frame of models, optionally quantile values,
and corresponding model coefficients
}}
}
\value{
rescaled_qra_fit object
}
\description{
Create a weighted_median_qra_fit object
}
