% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{par_grid_constructor_rel_wis_weighted_median}
\alias{par_grid_constructor_rel_wis_weighted_median}
\title{Parameter search grid constructor for rel_wis_weighted_median approach}
\usage{
par_grid_constructor_rel_wis_weighted_median(qfm_train, ...)
}
\arguments{
\item{qfm_train}{QuantileForecastMatrix with training set predictions from
component models}

\item{...}{mop up other arguments that are ignored}
}
\value{
vector of real-valued initial values for parameters
}
\description{
Parameter search grid constructor for rel_wis_weighted_median approach
}
