% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective_utils.R
\name{parse_model_case}
\alias{parse_model_case}
\title{function to extract model identifiers from abbreviation}
\usage{
parse_model_case(model_abbr)
}
\arguments{
\item{model_abbr}{abbreviation for a retrospective ensemble fit}
}
\value{
data frame of model fit characteristics
}
\description{
function to extract model identifiers from abbreviation
}
