% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecasts.R
\name{plot_forecasts_single_model}
\alias{plot_forecasts_single_model}
\title{plot forecasts}
\usage{
plot_forecasts_single_model(
  submissions_root,
  plots_root,
  forecast_date,
  model_abbrs,
  target_variables
)
}
\arguments{
\item{submissions_root}{path to folder with submission files, similar to
data-processed folder in covid19-forecast-hub repo}

\item{plots_root}{folder where plots will be saved}

\item{forecast_date}{date for Monday of model submissions to plot}

\item{model_abbrs}{model names to plot}

\item{target_variables}{character vector of variables like "cases", "deaths",
or "hospitalizations"}
}
\description{
plot forecasts
}
