% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{predict.qra_fit}
\alias{predict.qra_fit}
\title{Predict based on a quantile regression averaging fit}
\usage{
\method{predict}{qra_fit}(qra_fit, qfm)
}
\arguments{
\item{qra_fit}{object of class qra_fit}

\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}
}
\value{
object of class QuantileForecastMatrix with quantile forecasts
}
\description{
Predict based on a quantile regression averaging fit
}
