% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{predict.rescaled_qra_fit}
\alias{predict.rescaled_qra_fit}
\title{Predict based on a quantile regression averaging fit}
\usage{
\method{predict}{rescaled_qra_fit}(qra_fit, qfm, ...)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{...}{ignored additional arguments (in particular, absorbs the
sort_quantiles argument used by \code{predict.qra_fit})}

\item{qra_fit_rescalable}{object of class qra_fit_rescalable}
}
\value{
object of class QuantileForecastMatrix with quantile forecasts
}
\description{
Predict based on a quantile regression averaging fit
}
