% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{predict.weighted_median_qra_fit}
\alias{predict.weighted_median_qra_fit}
\title{Predict based on a quantile regression averaging fit}
\usage{
\method{predict}{weighted_median_qra_fit}(qra_fit, qfm, sort_quantiles = TRUE)
}
\arguments{
\item{qra_fit}{object of class weighted_median_qra_fit}

\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{sort_quantiles}{logical; if TRUE, the predictive quantiles are sorted
in order of the quantile level.  Otherwise, the raw predictive quantiles are
returned.}
}
\value{
object of class QuantileForecastMatrix with quantile forecasts
}
\description{
Predict based on a quantile regression averaging fit
}
