% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{print.qfm}
\alias{print.qfm}
\title{Print a QuantileForecastMatrix}
\usage{
\method{print}{qfm}(qfm)
}
\arguments{
\item{qfm}{wide matrix representation of forecasts}
}
\value{
a dataframe of all characters with no names
}
\description{
Print a QuantileForecastMatrix
}
