% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{qfm_data}
\alias{qfm_data}
\title{Show the basic data matrix of a QuantileForecastMatrix}
\usage{
qfm_data(qfm, digits = 1)
}
\arguments{
\item{qfm}{wide matrix representation of forecasts}

\item{digits}{number of digits to leave after decimal}
}
\description{
Show the basic data matrix of a QuantileForecastMatrix
}
