% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{quantile_score}
\alias{quantile_score}
\title{Calculate quantile scores for multiple quantile predictions at a single
probability level tau}
\usage{
quantile_score(y, q, tau)
}
\arguments{
\item{y}{vector of observations from target distribution}

\item{q}{predictive quantile values}

\item{tau}{the probability level for the predictive quantile q}
}
\value{
vector of quantile scores separately for each element of q
}
\description{
Calculate quantile scores for multiple quantile predictions at a single
probability level tau
}
