% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{softmax_matrix_rows}
\alias{softmax_matrix_rows}
\title{Calculate softmax for each row of a matrix}
\usage{
softmax_matrix_rows(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
softmax(X) by row; each row of result sums to 1
}
\description{
Calculate softmax for each row of a matrix
}
