% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{sort.QuantileForecastMatrix}
\alias{sort.QuantileForecastMatrix}
\title{Sort method for QuantileForecastMatrix class which
currently leaves any NA's encountered where they were
found, sorting model-wise around them}
\usage{
\method{sort}{QuantileForecastMatrix}(qfm)
}
\arguments{
\item{qfm}{wide matrix representation of forecasts}
}
\value{
a QFM with always increasing quantiles within given
model, target, and location
}
\description{
Sort method for QuantileForecastMatrix class which
currently leaves any NA's encountered where they were
found, sorting model-wise around them
}
