% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{str.QuantileForecastMatrix}
\alias{str.QuantileForecastMatrix}
\title{Convenience \code{str} method for QuantileForecastMatrix class which
avoids \code{seq_len(ncol(qfm))} errors and warnings}
\usage{
\method{str}{QuantileForecastMatrix}(qfm)
}
\arguments{
\item{qfm}{wide matrix representation of forecasts}
}
\value{
none
}
\description{
Convenience \code{str} method for QuantileForecastMatrix class which
avoids \code{seq_len(ncol(qfm))} errors and warnings
}
