% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{[.QuantileForecastMatrix}
\alias{[.QuantileForecastMatrix}
\title{Extract parts of a QuantileForecastMatrix object}
\usage{
\method{[}{QuantileForecastMatrix}(qfm, i = seq_len(nrow(qfm)), j = seq_len(ncol(qfm)), ...)
}
\arguments{
\item{qfm}{an object of class QuantileForecastMatrix}

\item{i}{integer indices of rows to extract}

\item{j}{integer indices of columns to extract}

\item{...}{other ignored arguments; used to mop up a required drop argument,
which is not supported.}
}
\value{
object of class QuantileForecastMatrix
}
\description{
Extract parts of a QuantileForecastMatrix object
}
