% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit_convenience.R
\name{train_test_split}
\alias{train_test_split}
\title{Split a QuantileForecastMatrix object into a training and test set}
\usage{
train_test_split(forecast_matrix, forecast_week_end_date, window_size)
}
\arguments{
\item{forecast_matrix}{a QuantileForecastMatrix with all train/test forecasts}

\item{forecast_week_end_date}{the forecast week end date defining the test
set}

\item{window_size}{integer number of weeks in the training set}
}
\value{
named list with entries qfm_train and qfm_test
}
\description{
\itemize{
\item if window_size >= 1, training set comprises only forecasts where
target_end_date <= forecast_week_end_date
\item else if window_size == 0, just keep horizon 1 for train set
This is not ideal behavior, but we need to rewrite the equal weighted
methods before we change this.
}
}
