% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{validate_qra_fit}
\alias{validate_qra_fit}
\title{Validate qra_fit object}
\usage{
validate_qra_fit(qra_fit)
}
\arguments{
\item{qra_fit}{}
}
\value{
invisible(TRUE) if qra_fit is valid;
otherwise, an error is thrown
}
\description{
Validate qra_fit object
}
