% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{validate_quantile_crossing}
\alias{validate_quantile_crossing}
\title{Validate quantile crossing}
\usage{
validate_quantile_crossing(qfm)
}
\arguments{
\item{qfm}{object of class QuantileForecastMatrix}
}
\value{
invisible(TRUE) if no quantile crossing; otherwise, error
}
\description{
Validate quantile crossing
}
