% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{validate_QuantileForecastMatrix}
\alias{validate_QuantileForecastMatrix}
\title{Validate QuantileForecastMatrix object}
\usage{
validate_QuantileForecastMatrix(qfm, strict = FALSE)
}
\arguments{
\item{qfm}{object of class QuantileForecastMatrix}

\item{strict}{logical; if FALSE (default), only checks that the arguments
satisfy minimum requirements of formatting; if TRUE, also checks
desirable properties of a quantile forecast such as quantile crossing}
}
\value{
invisible(TRUE) if QuantileForecastMatrix is valid;
otherwise, an error is thrown
}
\description{
Validate QuantileForecastMatrix object
}
