% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{validate_rescaled_qra_fit}
\alias{validate_rescaled_qra_fit}
\title{Validate rescaled_qra_fit object}
\usage{
validate_rescaled_qra_fit(rescaled_qra_fit)
}
\arguments{
\item{rescaled_qra_fit}{}
}
\value{
invisible(TRUE) if qra_fit is valid;
otherwise, an error is thrown
}
\description{
Validate rescaled_qra_fit object
}
