% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{weighted_interval_penalty}
\alias{weighted_interval_penalty}
\title{Calculate interval scores for multiple intervals at a single value of alpha}
\usage{
weighted_interval_penalty(y, l, u, alpha)
}
\arguments{
\item{y}{vector of observations from target distribution}

\item{l}{lower bound of predictive distribution}

\item{u}{upper bound of predictive distribution}

\item{alpha}{the interval \link{l,u} is a (1 - alpha)*100\% predictive interval}
}
\value{
vector of interval scores separately for each element of y
}
\description{
Calculate interval scores for multiple intervals at a single value of alpha
}
