% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{wis_loss}
\alias{wis_loss}
\title{Calculate wis_loss as a function of parameters used to construct a qra model}
\usage{
wis_loss(par, model_constructor, qfm_train, y_train)
}
\arguments{
\item{par}{real-valued vector of parameters}

\item{model_constructor}{a function that accepts a real-valued vector of
parameters and returns a model of class qra_fit}

\item{qfm_train}{QuantileForecastMatrix with training set predictions from
component models}

\item{y_train}{numeric vector of responses for training set}
}
\value{
scalar wis loss for given parameter values
}
\description{
Calculate wis_loss as a function of parameters used to construct a qra model
}
