% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{wis_loss_rel_wis_weights}
\alias{wis_loss_rel_wis_weights}
\title{Calculate wis_loss as a function of parameters used to construct a
relative WIS weighted qra model. This method is specific to the
relative WIS weighted mean and median methods, and allows for a
maximum weight; an infinite loss is returned if the largest weight
is greater than this value.}
\usage{
wis_loss_rel_wis_weights(
  par,
  model_constructor,
  qfm_train,
  y_train,
  max_weight = 1,
  ...
)
}
\arguments{
\item{par}{real-valued vector of parameters}

\item{model_constructor}{a function that accepts a real-valued vector of
parameters and returns a model of class qra_fit}

\item{qfm_train}{QuantileForecastMatrix with training set predictions from
component models}

\item{y_train}{numeric vector of responses for training set}

\item{max_weight}{numeric value for maximum weight. Defaults to 1}

\item{...}{arguments passed on to the model_constructor}
}
\value{
scalar wis loss for given parameter values
}
\description{
Calculate wis_loss as a function of parameters used to construct a
relative WIS weighted qra model. This method is specific to the
relative WIS weighted mean and median methods, and allows for a
maximum weight; an infinite loss is returned if the largest weight
is greater than this value.
}
