% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{wiw}
\alias{wiw}
\title{Calculate quantile score, optionally including absolute error,
for a parametric distribution.}
\usage{
wiw(y, qfm)
}
\arguments{
\item{y}{vector of observations from target distribution}

\item{qfm}{object of class QuantileForecastMatrix}
}
\value{
vector of interval scores separately for each element of x
}
\description{
Calculate quantile score, optionally including absolute error,
for a parametric distribution.
}
