% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eligibility.R
\name{calc_forecast_missingness}
\alias{calc_forecast_missingness}
\title{Compute forecast missingness for each combination of location and model}
\usage{
calc_forecast_missingness(qfm, by_target = FALSE, by_week = FALSE)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{by_target}{logical; if TRUE, results are returned by target, and if
FALSE results are summarized across targets.  by_target may only be set to
TRUE if by_week is TRUE.}

\item{by_week}{logical; if TRUE, results are returned by forecast week, and
if FALSE results are summarized across weeks.}
}
\value{
data frame with a row for each combination of
location, forecast week end date (if requestrd), and model and a logical column called
'any_missing' with entry TRUE if any forecasts were missing across all
quantiles and targets
}
\description{
Compute forecast missingness for each combination of location and model
}
