% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eligibility.R
\name{calc_forecast_missingness}
\alias{calc_forecast_missingness}
\title{Compute forecast missingness for each combination of unit and model}
\usage{
calc_forecast_missingness(qfm, lookback_length = 0, model_id_name)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{lookback_length}{non-negative integer number of historic weeks that
are examined for forecast missingness; 0 is appropriate for equal weight
ensembles where no historical data is required.  If two past weeks of
forecast data are required to estimate ensemble parameters, lookback_length
should be 2}

\item{model_id_name}{character name of column in forecast matrix col_index
specifying column}
}
\value{
data frame with a row for each combination of
unit, forecast week end date, and model and a logical column called
'any_missing' with entry TRUE if any forecasts were missing across all
quantiles and targets
}
\description{
Compute forecast missingness for each combination of unit and model
}
