% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eligibility.R
\name{calc_model_eligibility_for_ensemble}
\alias{calc_model_eligibility_for_ensemble}
\title{Generate a data frame with a row for each model and an indicator of whether
that model is eligible for inclusion in an ensemble, with explanation if not}
\usage{
calc_model_eligibility_for_ensemble(
  qfm,
  observed_by_unit_target_end_date,
  lookback_length
)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{observed_by_unit_target_end_date}{data frame of observed values}

\item{lookback_length}{non-negative integer number of historic weeks that
are examined for forecast missingness; 0 is appropriate for equal weight
ensembles where no historical data is required.  If two past weeks of
forecast data are required to estimate ensemble parameters, lookback_length
should be 2}
}
\value{
data frame with two columns:
\itemize{
\item one with name given by model_id_name recording model id for each model
\item a second called 'eligibility' that is either the string 'eligible' or a
brief description of why the model can't be included
}
}
\description{
Generate a data frame with a row for each model and an indicator of whether
that model is eligible for inclusion in an ensemble, with explanation if not
}
