% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{estimate_qra}
\alias{estimate_qra}
\title{Estimate Quantile Regression Averaging model}
\usage{
estimate_qra(
  qfm_train,
  y_train,
  qfm_test = NULL,
  qra_model = c("ew", "convex_per_model", "unconstrained_per_model",
    "rescaled_convex_per_model"),
  quantile_groups = NULL,
  backend = "optim",
  ...
)
}
\arguments{
\item{qfm_train}{QuantileForecastMatrix with training set predictions}

\item{y_train}{numeric vector of responses for training set}

\item{qfm_test}{QuantileForecastMatrix with test set predictions}

\item{qra_model}{quantile averaging model: currently only 'ew' is supported}

\item{quantile_groups}{Vector of group labels for quantiles, having the same
length as the number of quantiles.  Common labels indicate that the ensemble
weights for the corresponding quantile levels should be tied together.
Default is rep(1,length(quantiles)), which means that a common set of
ensemble weights should be used across all levels.  This is the argument
\code{tau_groups} for \code{quantmod::quantile_ensemble}, and may only be supplied if
\verb{backend = 'quantmod}}

\item{backend}{implementation used for estimation; currently either
'optim', using L-BFGS-B as provided by the optim function in R;
'NlcOptim', using NlcOptim::solnl; or 'quantmod', using
quantmod::quantile_ensemble}
}
\value{
object of class qra_fit
}
\description{
Estimate Quantile Regression Averaging model
}
