% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{estimate_qra}
\alias{estimate_qra}
\title{Estimate Quantile Regression Averaging model}
\usage{
estimate_qra(
  qfm_train,
  y_train,
  qfm_test = NULL,
  intercept = FALSE,
  constraint = c("ew", "convex", "positive", "unconstrained"),
  quantile_groups = NULL,
  backend = "optim",
  ...
)
}
\arguments{
\item{qfm_train}{QuantileForecastMatrix with training set predictions}

\item{y_train}{numeric vector of responses for training set}

\item{qfm_test}{QuantileForecastMatrix with test set predictions}

\item{intercept}{logical specifying whether an intercept is included}

\item{constraint}{character specifying constraints on parameters; 'ew',
'convex', 'positive' or 'unconstrained'}

\item{quantile_groups}{Vector of group labels for quantiles, having the same
length as the number of quantiles.  Common labels indicate that the ensemble
weights for the corresponding quantile levels should be tied together.
Default is rep(1,length(quantiles)), which means that a common set of
ensemble weights should be used across all levels.  This is the argument
\code{tau_groups} for \code{quantgen::quantile_ensemble}, and may only be supplied if
\verb{backend = 'quantgen}}

\item{backend}{implementation used for estimation; currently either
'optim', using L-BFGS-B as provided by the optim function in R;
'NlcOptim', using NlcOptim::solnl; or 'quantgen', using
quantgen::quantile_ensemble}
}
\value{
object of class qra_fit
}
\description{
Estimate Quantile Regression Averaging model
}
