% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{estimate_qra_optimized}
\alias{estimate_qra_optimized}
\title{Estimate qra parameters by optimizing weighted interval score loss}
\usage{
estimate_qra_optimized(
  qfm_train,
  y_train,
  quantile_groups,
  qra_model = c("convex_per_model", "unconstrained_per_model",
    "rescaled_convex_per_model", "rel_wis_weighted_median"),
  backend = c("optim", "NlcOptim", "grid_search"),
  max_weight = NULL
)
}
\arguments{
\item{qfm_train}{QuantileForecastMatrix with training set predictions from
component models}

\item{y_train}{numeric vector of responses for training set}

\item{quantile_groups}{Vector of group labels for quantiles, having the same
length as the number of quantiles.  Common labels indicate that the ensemble
weights for the corresponding quantile levels should be tied together.
Default is rep(1,length(quantiles)), which means that a common set of
ensemble weights should be used across all levels.  This is the argument
\code{tau_groups} for \code{quantgen::quantile_ensemble}, and may only be supplied if
\verb{backend = 'quantgen}}

\item{qra_model}{quantile averaging model}

\item{backend}{implementation used for estimation; currently either
'optim', using L-BFGS-B as provided by the optim function in R, or
'NlcOptim', using NlcOptim::solnl, or 'grid_search', using
covidEnsemble::grid_search}

\item{max_weight}{numeric value for maximum weight. Ignored unless qra_model
is rel_wis_weighted_median or rel_wis_weighted mean and backend is
grid_search}
}
\value{
object of class qra_fit
}
\description{
Estimate qra parameters by optimizing weighted interval score loss
}
