% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{extract_weights_qenspy_qra_fit}
\alias{extract_weights_qenspy_qra_fit}
\title{Predict based on a quantile regression averaging fit}
\usage{
extract_weights_qenspy_qra_fit(qra_fit)
}
\arguments{
\item{qra_fit}{object of class qra_fit}
}
\value{
object of class QuantileForecastMatrix with quantile forecasts
}
\description{
Predict based on a quantile regression averaging fit
}
