% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{new_QuantileForecastMatrix_from_df}
\alias{new_QuantileForecastMatrix_from_df}
\title{Translate from a tidy data frame of forecasts to a wide matrix representation
suitable for use by \code{calc_qra_from_matrices}}
\usage{
new_QuantileForecastMatrix_from_df(
  forecast_df,
  model_col,
  id_cols,
  quantile_name_col,
  quantile_value_col
)
}
\arguments{
\item{forecast_df}{data frame of forecasts from component models in 'CDC format'}

\item{model_col}{character name of column in forecast_df identifying model}

\item{id_cols}{character vector of columns in forecast_df identifying unique forecast
settings such as location, time zero, and horizon}

\item{quantile_name_col}{character name of column in forecast_df containing probability
for quantile, e.g. 0.95}

\item{quantile_value_col}{character name of column in forecast_df identifying value for
quantile, e.g. 195}
}
\value{
Object of class QuantileForecastMatrix
}
\description{
Translate from a tidy data frame of forecasts to a wide matrix representation
suitable for use by \code{calc_qra_from_matrices}
}
