% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra_fit.R
\name{predict.rescaled_qra_fit}
\alias{predict.rescaled_qra_fit}
\title{Predict based on a rescaled quantile regression averaging fit}
\usage{
\method{predict}{rescaled_qra_fit}(qra_fit, qfm, sort_quantiles = TRUE, ...)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{sort_quantiles}{logical; if TRUE, the predictive quantiles are sorted
in order of the quantile level.  Otherwise, the raw predictive quantiles are
returned.}

\item{...}{ignored additional arguments}

\item{qra_fit_rescalable}{object of class qra_fit_rescalable}
}
\value{
object of class QuantileForecastMatrix with quantile forecasts
}
\description{
Predict based on a rescaled quantile regression averaging fit
}
