% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileForecastMatrix.R
\name{validate_QuantileForecastMatrix}
\alias{validate_QuantileForecastMatrix}
\title{Validate QuantileForecastMatrix object}
\usage{
validate_QuantileForecastMatrix(qfm)
}
\arguments{
\item{qfm}{object of class QuantileForecastMatrix}
}
\value{
invisible(TRUE) if ncopula is valid; otherwise, an error is thrown
}
\description{
Validate QuantileForecastMatrix object
}
