% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{wis}
\alias{wis}
\title{Calculate quantile score, optionally including absolute error,
for a parametric distribution.}
\usage{
wis(y, qfm, quantile_probs)
}
\arguments{
\item{y}{vector of observations from target distribution}

\item{alphas}{vector of alphas.  the score is based on several
(1 - alpha)*100\% predictive intervals for each alpha in alphas}

\item{ae}{logical.  If true, include absolute error in score}

\item{family}{family for distribution; e.g. 'norm' or 'nbinom'}

\item{args}{list of arguments for quantile function of parametric family,
e.g. list(mean=0.0, sd=1.0)}
}
\value{
vector of interval scores separately for each element of x
}
\description{
Calculate quantile score, optionally including absolute error,
for a parametric distribution.
}
